fix message service test
